+++ Riot/Managers/Widgets/WidgetManager.h
@@ -56,6 +56,7 @@ typedef enum : NSUInteger
     WidgetManagerErrorCodeNotEnoughPower,
     WidgetManagerErrorCodeCreationFailed,
     WidgetManagerErrorCodeNoIntegrationsServerConfigured,
+    WidgetManagerErrorCodeDisabledIntegrationsServer,
     WidgetManagerErrorCodeFailedToConnectToIntegrationsServer,
     WidgetManagerErrorCodeTermsNotSigned
 }